+++ src/domain/sensor_domain/processors/i_reading_processor.h
@@ -1,12 +1,14 @@
 #pragma once
 
+#include <cstdint>
+
 namespace eerie_leap::domain::sensor_domain::processors {
 
 class IReadingProcessor {
 public:
     virtual ~IReadingProcessor() = default;
 
-    virtual void ProcessReading(const size_t sensor_id_hash) = 0;
+    virtual void ProcessReading(const uint32_t sensor_id_hash) = 0;
 };
 
 } // namespace eerie_leap::domain::sensor_domain::processors
